+++ src/translate/printf.ts
@@ -1,7 +1,7 @@
 import { formatValues } from '../localize/formatValues';
 import { Formatter } from '../types/StatDescription';
 
-export type Params = number[];
+export type Params = Array<number | [number, number]>;
 type PreparedParams = Array<number | string>;
 
 export default function printf(