+++ src/types/StatValue.ts
@@ -1,0 +1,5 @@
+export type StatValue = number | Range;
+export type Range = [number, number];
+
+export const isRange = (value: StatValue): value is Range =>
+  Array.isArray(value) && value.length === 2;