+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/eps1lon/poe-mods/compare/v0.2.0...HEAD)
+### Added
+- `formatStats` now supports a range as a stat value. Check out the API documentation
+  for further information.
 
 ## [0.2.0](https://github.com/eps1lon/poe-mods/compare/v0.1.3...v0.2.0) (2017-20-03)
 ### Added